#18 REFACTOR: 패키지 구조를 재배치한다
