+++ src/main/java/com/project/zighang/global/service/OcrReader.java
@@ -1,0 +1,11 @@
+package com.project.zighang.global.service;
+
+import com.project.zighang.global.dto.ClovaOcrResponse;
+
+public interface OcrReader {
+    ClovaOcrResponse extractFromUrl(String imageUrl);
+
+    String extractSuccessTxtFromUrl(String imageUrl);
+
+    String extractTxtFromUrl(String imageUrl);
+}
